# src/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ChainInfoResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"chains": {
						"items": {
							"properties": {
								"rpcUrl": {
									"type": "string"
								},
								"explorerUrl": {
									"type": "string"
								},
								"currency": {
									"type": "string"
								},
								"testnet": {
									"type": "boolean"
								},
								"name": {
									"type": "string"
								},
								"chainId": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"rpcUrl",
								"explorerUrl",
								"currency",
								"testnet",
								"name",
								"chainId"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"chains"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SupportedChain": {
				"properties": {
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"testnet": {
						"type": "boolean"
					},
					"currency": {
						"type": "string"
					},
					"explorerUrl": {
						"type": "string"
					},
					"rpcUrl": {
						"type": "string"
					},
					"nativeCurrency": {
						"properties": {
							"decimals": {
								"type": "number",
								"format": "double"
							},
							"symbol": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"decimals",
							"symbol",
							"name"
						],
						"type": "object"
					}
				},
				"required": [
					"chainId",
					"name",
					"testnet",
					"currency",
					"explorerUrl",
					"rpcUrl",
					"nativeCurrency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SupportedChainsResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"chains": {
						"items": {
							"$ref": "#/components/schemas/SupportedChain"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"chains"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NativeBalance": {
				"properties": {
					"balance": {
						"type": "string"
					},
					"formatted": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"decimals": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"balance",
					"formatted",
					"currency",
					"decimals"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BalanceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"walletAddress": {
						"type": "string"
					},
					"nativeBalance": {
						"$ref": "#/components/schemas/NativeBalance"
					},
					"chainId": {
						"type": "number",
						"format": "double"
					},
					"chainName": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"walletAddress",
					"nativeBalance",
					"chainId",
					"chainName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"app": {
				"type": "apiKey",
				"name": "X-App-ID",
				"in": "header",
				"description": "Application ID for API access"
			},
			"bearer": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT token for authentication"
			}
		}
	},
	"info": {
		"title": "crefy-connect-backend",
		"version": "1.0.0",
		"description": "Wallet authentication API with Express and TypeScript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Jimleston Osoi"
		}
	},
	"paths": {
...
		"/signing/chains": {
			"get": {
				"operationId": "GetSupportedChains",
				"responses": {
					"200": {
						"description": "Supported chains retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChainInfoResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"chains": [
												{
													"chainId": 1,
													"name": "Ethereum",
													"testnet": false,
													"currency": "ETH",
													"explorerUrl": "https://etherscan.io",
													"rpcUrl": "https://eth.llamarpc.com"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all supported chains",
				"tags": [
					"Signing Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
...
		"/balance/chains": {
			"get": {
				"operationId": "GetSupportedChains",
				"responses": {
					"200": {
						"description": "Supported chains retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SupportedChainsResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"chains": [
												{
													"chainId": 8453,
													"name": "Base",
													"testnet": false,
													"currency": "ETH",
													"explorerUrl": "https://basescan.org",
													"rpcUrl": "https://mainnet.base.org",
													"nativeCurrency": {
														"name": "Ethereum",
														"symbol": "ETH",
														"decimals": 18
													}
												},
												{
													"chainId": 11155111,
													"name": "Sepolia",
													"testnet": true,
													"currency": "ETH",
													"explorerUrl": "https://sepolia.etherscan.io",
													"rpcUrl": "https://sepolia.infura.io/v3/",
													"nativeCurrency": {
														"name": "Sepolia ETH",
														"symbol": "ETH",
														"decimals": 18
													}
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get all supported chains for balance checking",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/balance/native": {
			"get": {
				"operationId": "GetBalance",
				"responses": {
					"200": {
						"description": "Native balance retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BalanceResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"walletAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"nativeBalance": {
												"balance": "125000000000000000",
												"formatted": "0.125",
												"currency": "ETH",
												"decimals": 18
											},
											"chainId": 8453,
											"chainName": "Base"
										}
									}
								}
							}
						}
					}
				},
				"description": "Get native balance for a wallet address on a specific chain",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "chainId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/balance/balances": {
			"get": {
				"operationId": "GetBalances",
				"responses": {
					"200": {
						"description": "Balances retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"balances": {
											"items": {
												"properties": {
													"chainName": {
														"type": "string"
													},
													"decimals": {
														"type": "number",
														"format": "double"
													},
													"currency": {
														"type": "string"
													},
													"chainId": {
														"type": "number",
														"format": "double"
													},
													"formatted": {
														"type": "string"
													},
													"balance": {
														"type": "string"
													}
												},
												"required": [
													"chainName",
													"decimals",
													"currency",
													"chainId",
													"formatted",
													"balance"
												],
												"type": "object"
											},
											"type": "array"
										},
										"walletAddress": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"balances",
										"walletAddress",
										"success"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"success": true,
											"walletAddress": "0xa5E0Da329eE5AA03f09228e534953496334080f5",
											"balances": [
												{
													"balance": "125000000000000000",
													"formatted": "0.125",
													"chainId": 8453,
													"currency": "ETH",
													"decimals": 18,
													"chainName": "Base"
												},
												{
													"balance": "50000000000000000",
													"formatted": "0.05",
													"chainId": 11155111,
													"currency": "ETH",
													"decimals": 18,
													"chainName": "Sepolia"
												}
											]
										}
									}
								}
							}
						}
					}
				},
				"description": "Get native balances across multiple chains",
				"tags": [
					"Balance Service"
				],
				"security": [
					{
						"app": []
					},
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "chainIds",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "network",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}